Move activity starter to internal interface
